refactor: separate queries and mutations
